Move MissionPages component from pages directory to components
